funcionalidad estrellas, ajustes card estrellas: letra, tamanios ubicacion, json calificaciones,

+++ baricharafriends/src/main/java/com/edw88/baricharafriends/model/SitiosItem.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 data class SitiosItem(
     @SerializedName("calificacion")
-    val calificacion: Double,
+    val calificacion: Float,
     @SerializedName("descripcion")
     val descripcion: String,
     @SerializedName("nombre")
